Copy files when attaching, code cleanup

+++ astrid/plugin-src/com/todoroo/astrid/files/FilesControlSet.java
@@ -124,10 +124,13 @@ public class FilesControlSet extends PopupControlSet {
                             new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface d, int which) {
-                            metadataService.delete(m);
-                            files.remove(m);
-                            refreshDisplayView();
-                            finalList.removeView(fileRow);
+                            File f = new File(m.getValue(FileMetadata.FILE_PATH));
+                            if (f.delete()) {
+                                metadataService.delete(m);
+                                files.remove(m);
+                                refreshDisplayView();
+                                finalList.removeView(fileRow);
+                            }
                         }
                     }, null);
                 }
